requirements.txt
flask==3.0.1
line-bot-sdk==3.9.0
sqlalchemy==2.0.25
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
